Correction stop script.
In case the app is not running, simply exit 0.

+++ aws-bin/app-stop.sh
@@ -1,3 +1,7 @@
 #!/usr/bin/env bash
 
-kill `pgrep -f '@angular/cli'` > /home/ec2-user/app-stop.log
+if pgrep -f '@angular/cli'; then
+  kill `pgrep -f '@angular/cli'` > /home/ec2-user/app-stop.log
+fi
+
+exit 0
